requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
aiogram==3.*
